+++ src/Network/Network/AzureFirewall/ApplicationRule/NewAzureFirewallApplicationRuleCommand.cs
@@ -47,6 +47,12 @@ namespace Microsoft.Azure.Commands.Network
         public string[] SourceAddress { get; set; }
 
         [Parameter(
+            Mandatory = false,
+            HelpMessage = "The source ipgroup of the rule")]
+        [ValidateNotNullOrEmpty]
+        public string[] SourceIpGroup { get; set; }
+
+        [Parameter(
             Mandatory = true,
             ParameterSetName = AzureFirewallApplicationRuleParameterSets.TargetFqdn,
             HelpMessage = "The target FQDNs of the rule")]
@@ -77,6 +83,8 @@ namespace Microsoft.Azure.Commands.Network
                 FqdnTag = AzureFirewallFqdnTagHelper.MapUserInputToAllowedFqdnTags(FqdnTag, this.AzureFirewallFqdnTagClient).ToArray();
             }
 
+            Array.ForEach(this.SourceIpGroup,r => ValidateIsIpGroup(r));
+
             var protocolsAsWeExpectThem = MapUserProtocolsToFirewallProtocols(Protocol?.ToList());
 
             var applicationRule = new PSAzureFirewallApplicationRule
@@ -84,6 +92,7 @@ namespace Microsoft.Azure.Commands.Network
                 Name = this.Name,
                 Description = this.Description,
                 SourceAddresses = this.SourceAddress?.ToList(),
+                SourceIpGroups = this.SourceIpGroup?.ToList(),
                 Protocols = protocolsAsWeExpectThem,
                 TargetFqdns = this.TargetFqdn?.ToList(),
                 FqdnTags = this.FqdnTag?.ToList()