+++ src/Network/Network/ChangeLog.md
@@ -20,6 +20,7 @@
 
 ## Upcoming Release
 * Change `Start-AzVirtualNetworkGatewayConnectionPacketCapture.md` and `Start-AzVirtualnetworkGatewayPacketCapture.md` FilterData option examples.
+* Support for IpGroups in AzureFirewall Application,Nat & Network Rules.
 
 ## Version 2.0.0
 * Change all cmdlets for PrivateEndpointConnection to support generic service provider.